+++ packages/boustro/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased]
+
+- **Changed**: Boustro now uses marker characters to detect backspace on empty lines.
+  See [this blog post](https://medium.com/super-declarative/why-you-cant-detect-a-delete-action-in-an-empty-flutter-text-field-3cf53e47b631) for an explanation on why that's necessary.
+
 ## [0.1.1]
 
 - **Added** textSelectable flag to DocumentView.